aula :: value notifier
